+++ include/geometry/box.hpp
@@ -9,9 +9,9 @@ class Box : public Geometry
 private:
     TriangleMesh _mesh;
 public:
-    Box(Geometric center, std::array<float,3> sides, std::array<Geometric,3> axis);
+    Box(Geometric center, std::array<double,3> sides, std::array<Geometric,3> axis);
 
-    float implicit(Geometric x) const;
+    double implicit(Geometric x) const;
 
     bool intersect_with_ray(const Ray& r, Intersection& intersection) const;
 };
