+++ include/geometry/sphere.hpp
@@ -6,11 +6,11 @@ class Sphere : Geometry
 {
 private:
     Geometric _center;
-    float _radius;
+    double _radius;
 public:
-    Sphere(Geometric center, float radius);
+    Sphere(Geometric center, double radius);
     
-    float implicit(Geometric x) const;
+    double implicit(Geometric x) const;
 
     bool intersect_with_ray(const Ray& r, Intersection& intersection) const;
 };
