+++ src/intersection.cpp
@@ -1,17 +1,17 @@
 #include "intersection.hpp"
 
-Intersection::Intersection(float distance, Geometric normal, Geometric point):
+Intersection::Intersection(double distance, Geometric normal, Geometric point):
     _distance(distance),_normal(normal),_point(point){}
 
 Intersection::Intersection():
-    _distance(std::numeric_limits<float>::max()),_normal(Geometric::vector0()),_point(Geometric::point0()){}
+    _distance(std::numeric_limits<double>::max()),_normal(Geometric::vector0()),_point(Geometric::point0()){}
 
 Geometric Intersection::point() const
 {
     return this->_point;
 }
 
-float Intersection::distance() const
+double Intersection::distance() const
 {
     return this->_distance;
 }
